Fix size of EEPROM tmc_hybrid_threshold
Followup to #10101

+++ Marlin/configuration_store.cpp
@@ -1379,7 +1379,7 @@ void MarlinSettings::postprocess() {
           #endif
         }
       #else
-        uint16_t thrs_val;
+        uint32_t thrs_val;
         for (uint8_t q=TMC_AXES; q--;) EEPROM_READ(thrs_val);
       #endif
